gyro
The panorama function needs to be removed. So the texture doesn't matter. In the test, the texture of the classroom is not used, so it does not look like a classroom. Models that will be added to the classroom at a later stage. Only having a gyroscope to adjust the viewing angle is useful at this stage